test_nta_api.c: silenced warning when printf()ing pid_t.
darcs-hash:20060516134347-65a35-3e32f1120c6f1611fb359d09c4b0e33f700c3d6d.gz

+++ libsofia-sip-ua/nta/test_nta_api.c
@@ -1299,7 +1299,7 @@ int main(int argc, char *argv[])
 
   if (o_attach) {
     char line[10];
-    printf("nua_test: pid %u\n", getpid());
+    printf("nua_test: pid %lu\n", (unsigned long)getpid());
     printf("<Press RETURN to continue>\n");
     fgets(line, sizeof line, stdin);
   }
